TLS 1.2: Factor out common key exchange decoding logic and error reporting.
Move the TLS-1.2-specific key agreement logic out of the general-purpose `kx`
module and into a new `crate::tls12` module that can hold common logic for
clients and servers.

Ensure error reporting and alerting for key agreement decoding are handled
consistently between client and server by consolidating the common logic.
